added Barrier events
