+++ GW2EIEvtcParser/Extensions/ExtensionCombatData/ExtensionCombatEvents/HealingStats/EXTAbstractBarrierEvent.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using GW2EIEvtcParser.ParsedData;
+using static GW2EIEvtcParser.Extensions.HealingStatsExtensionHandler;
+
+namespace GW2EIEvtcParser.Extensions
+{
+    public abstract class EXTAbstractHealingEvent : EXTAbstractHealingExtensionEvent
+    {
+        public int HealingDone { get; protected set; }
+
+        internal EXTAbstractHealingEvent(CombatItem evtcItem, AgentData agentData, SkillData skillData) : base(evtcItem, agentData, skillData)
+        {
+        }
+
+        public EXTHealingType GetHealingType(ParsedEvtcLog log)
+        {
+            return log.CombatData.EXTHealingCombatData.GetHealingType(Skill, log);
+        }
+
+    }
+}
