+++ GW2EIEvtcParser/Extensions/ExtensionCombatData/ExtensionCombatEvents/HealingStats/EXTNonDirectBarrierEvent.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using GW2EIEvtcParser.ParsedData;
+
+namespace GW2EIEvtcParser.Extensions
+{
+    public class EXTNonDirectBarrierEvent : EXTAbstractBarrierEvent
+    {
+
+        internal EXTNonDirectBarrierEvent(CombatItem evtcItem, AgentData agentData, SkillData skillData) : base(evtcItem, agentData, skillData)
+        {
+            BarrierGiven = -evtcItem.BuffDmg;
+            AgainstDowned = evtcItem.Pad1 == 1;
+        }
+    }
+}
